[errmapgen] Extract ErrorMapping from errmapgen to move-core-types
We need to pull out the `ErrorMapping` definitions so they're separated
from the generation logic, which depends on the internal, unpublished
`move-model` crate. This separation will allow us to easily read and use
errmap's inside code from other published crates.
