+++ language/move-core/types/src/lib.rs
@@ -5,6 +5,7 @@
 
 pub mod account_address;
 pub mod effects;
+pub mod errmap;
 pub mod gas_schedule;
 pub mod identifier;
 pub mod language_storage;
