+++ language/tools/move-explain/Cargo.toml
@@ -13,7 +13,6 @@ edition = "2018"
 structopt = "0.3.21"
 diem-framework-releases = { path = "../../diem-framework/releases" }
 diem-workspace-hack = { path = "../../../common/workspace-hack" }
-errmapgen = { path = "../../move-prover/errmapgen" }
 move-core-types = { path = "../../move-core/types" }
 bcs = "0.1.2"
